Fix confusion with duplicate names
When a kind reports the same name for its singular and its plural
version then kubectl-tree gets confused and there's no way for a user
to query the tree for that specific kind.

With this commit, kubectl-tree eliminates duplicates in its internal
list of names for a particular kind so that there's always an
unambiguous way to specify which kind to query.

closes #64

Signed-off-by: Max Jonas Werner <mail@makk.es>